+++ test/PluginTest.php
@@ -1,14 +1,15 @@
 <?php
+
 /**
- * @link      http://github.com/zfcampus/zf-asset-manager for the canonical source repository
- * @copyright Copyright (c) 2016 Zend Technologies USA Inc. (http://www.zend.com)
- * @license   http://framework.zend.com/license/new-bsd New BSD License
+ * @see       https://github.com/laminas-api-tools/api-tools-asset-manager for the canonical source repository
+ * @copyright https://github.com/laminas-api-tools/api-tools-asset-manager/blob/master/COPYRIGHT.md
+ * @license   https://github.com/laminas-api-tools/api-tools-asset-manager/blob/master/LICENSE.md New BSD License
  */
 
-namespace ZFTest\AssetManager;
+namespace LaminasTest\ApiTools\AssetManager;
 
+use Laminas\ApiTools\AssetManager\Plugin;
 use PHPUnit_Framework_TestCase as TestCase;
-use ZF\AssetManager\Plugin;
 
 class PluginTest extends TestCase
 {