+++ CHANGELOG.md
@@ -18,7 +18,7 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Fixed
 
-- [#3](https://github.com/zfcampus/zf-asset-manager/pull/3) fixes the
+- [zfcampus/zf-asset-manager#3](https://github.com/zfcampus/zf-asset-manager/pull/3) fixes the
   `onPostPackageUpdate()` logic to pull both the initial and target package from
   the `UpdateOperation` composed in the `PackageEvent`, and to create
   new `PackageEvent` instances containing appropriate `UninstallOperation` and
@@ -28,7 +28,7 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Added
 
-- [#2](https://github.com/zfcampus/zf-asset-manager/pull/2) adds a
+- [zfcampus/zf-asset-manager#2](https://github.com/zfcampus/zf-asset-manager/pull/2) adds a
   post-package-update event handler that uninstalls assets for the package, and
   then installs any defined in the new package version. This feature will allow
   seamless updating of assets as they are updated in package dependencies.
@@ -61,7 +61,7 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Fixed
 
-- [#1](https://github.com/zfcampus/zf-asset-manager/pull/1) fixes how the
+- [zfcampus/zf-asset-manager#1](https://github.com/zfcampus/zf-asset-manager/pull/1) fixes how the
   `public/.gitignore` file is populated, ensuring no duplicates are created.
 
 ## 1.0.0 - 2016-07-26